refactor: review-comment: Inline function to improve typing
