+++ src/Frontend/state/actions/user-settings-actions/user-settings-actions.ts
@@ -3,8 +3,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 import { UserSettings } from '../../../../shared/shared-types';
-import { State } from '../../../types/types';
-import { getUserSettings } from '../../selectors/user-settings-selector';
 import { AppThunkAction } from '../../types';
 import { ACTION_SET_USER_SETTING, SetUserSetting } from './types';
 
@@ -19,30 +17,5 @@ export function fetchUserSettings(): AppThunkAction {
   return async (dispatch) => {
     const userSettings = await window.electronAPI.getUserSettings();
     dispatch(setUserSetting(userSettings));
-  };
-}
-
-function getUserSettingsToSet(
-  userSettings:
-    | Partial<UserSettings>
-    | ((currentSettings: UserSettings) => Partial<UserSettings>),
-  getState: () => State,
-): Partial<UserSettings> {
-  if (typeof userSettings === 'function') {
-    const currentUserSettings = getUserSettings(getState());
-    return userSettings(currentUserSettings);
-  }
-  return userSettings;
-}
-
-export function updateUserSettings(
-  userSettings:
-    | Partial<UserSettings>
-    | ((currentSettings: UserSettings) => Partial<UserSettings>),
-): AppThunkAction {
-  return async (dispatch, getState) => {
-    const userSettingsToSet = getUserSettingsToSet(userSettings, getState);
-    await window.electronAPI.updateUserSettings(userSettingsToSet);
-    dispatch(setUserSetting(userSettingsToSet));
   };
 }
